+++ tests/behat/behat_mod_unilabel.php
@@ -25,6 +25,7 @@
  */
 
 // NOTE: no MOODLE_INTERNAL test here, this file may be required by behat before including /config.php.
+use Behat\Gherkin\Node\TableNode as TableNode;
 
 require_once(__DIR__ . '/../../../../lib/behat/behat_base.php');
 
@@ -35,4 +36,49 @@ require_once(__DIR__ . '/../../../../lib/behat/behat_base.php');
  * @license   http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 class behat_mod_unilabel extends behat_base {
+    /**
+     * Adds the selected activity/resource filling the form data with the specified field/value pairs.
+     *
+     * Sections 0 and 1 are also allowed on frontpage.
+     *
+     * @Given I add a unilabel to course :coursefullname section :sectionnum and I fill the form with:
+     * @param string $coursefullname The course full name of the course.
+     * @param int $section The section number
+     * @param TableNode $data The activity field/value data
+     */
+    public function i_add_unilabel_and_i_fill_the_form_with($coursefullname, $section, TableNode $data) {
+
+        // Add activity to section.
+        $this->execute(
+            "behat_mod_unilabel::i_add_unilabel_to_course_section",
+            [$this->escape($coursefullname), $this->escape($section)]
+        );
+
+        // Wait to be redirected.
+        $this->execute('behat_general::wait_until_the_page_is_ready');
+
+        // Set form fields.
+        $this->execute("behat_forms::i_set_the_following_fields_to_these_values", $data);
+
+        // Save course settings.
+        $this->execute("behat_forms::press_button", get_string('savechangesandreturntocourse'));
+    }
+
+    /**
+     * Open a add activity form page.
+     *
+     * @Given I add a unilabel to course :coursefullname section :sectionnum
+     * @throws coding_exception
+     * @param string $coursefullname The course full name of the course.
+     * @param string $sectionnum The section number.
+     */
+    public function i_add_unilabel_to_course_section(string $coursefullname, string $sectionnum): void {
+        $addurl = new moodle_url('/course/modedit.php', [
+            'add' => 'unilabel',
+            'course' => $this->get_course_id($coursefullname),
+            'section' => intval($sectionnum),
+        ]);
+        $this->execute('behat_general::i_visit', [$addurl]);
+    }
+
 }