+++ tests/lib_test.php
@@ -35,14 +35,14 @@ namespace mod_unilabel;
  * @copyright   2018 onwards Grabs EDV {@link https://www.grabs-edv.de}
  * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
-class lib_test extends \advanced_testcase {
+final class lib_test extends \advanced_testcase {
     /**
      * Test create an instance.
      *
      * @covers \mod_unilabel\factory
      * @return void
      */
-    public function test_get_only_active_unilabeltypes() {
+    public function test_get_only_active_unilabeltypes(): void {
         global $DB;
         $this->resetAfterTest();
         $this->setAdminUser();